Adds migrate and upgrade commands
